[cf] Make use of Config* types from decl library.
The fuchsia.component.config library is being removed, this
change begins that migration.

The new types are ABI-compatible with the old ones, so as
long as everything builds this change should be a no-op.

The RealmBuilder protocol is not updated because of difficulty
with OOT versioning of the Dart client library, and this change
adds conversion functions to cm_rust to allow the Rust client
library for RB to interoperate with other libraries.

Bug: 96254
Change-Id: Id0e9ee3edef215397c28731ab6f16a2096e2bbb2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/850817
API-Review: Gary Bressler <geb@google.com>
Fuchsia-Auto-Submit: Adam Perry <adamperry@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Laura Peskin <pesk@google.com>