[paver] Support Fxfs images
This adds a new flag to bootserver, so clients wishing to use Fxblob
wil have to rebuild their bootserver after this patch.  Since Fxblob
isn't on by default, this is not a breaking change.

We will need to bump the netboot version when enabling Fxblob, as the
paver needs some new functionality for these images.

Bug: 122056
Change-Id: Id7f9669ac8a38583d14a9462302446fa1e32b45a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/847001
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Stephen Demos <sdemos@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>