[python] Add excludes to pyrightconfig.json
When I open a python file in fuchsia.git, pyright tries to index the
entire workspace, which for me is 351403 python files. It then pins a
core for the forseeable future and remains generally unresponsive until
it happens to index the file that I'm working in.

By excluding prebuilts, third_party, and build dirs the size of the
workspace goes down to 547 python files, and pyright can finish a
complete workspace check in under 15 seconds.

Change-Id: I23a4f6c8225f7d10ecb93d3928e6126a11641a14
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/851543
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Joseph Ryan <josephry@google.com>