[zither][c] Generate README.md
We update the C backend to generate a README.md comprised of
libray-level FIDL documentation. We only need this for the C backend at
this point, but this support can trivially be added to other backends
if/when we care to.

Bug: 111453
Change-Id: Ifd3aa568bd019076540f68729d9cf58447c78deb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/851998
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>