[superproject] Roll llvm-project/libc [libc] Fix hanging test on NVPTX due to lack of warp sync
Previously this wasn't implemented because it's effectively a no-op.
However, this should be safe to emit on sm_60 architectures. It's
important because it carries semantic importance for whether or not
something can be moved. So we should always emit this instrinsic.

Differential Revision: https://reviews.llvm.org/D149923

GitOrigin-RevId: bc1aef6b1b718adc821b24f50c2d7b6c67c8e82d
Original-Revision: 2f6056949920ff3667294b14ebcb0074a2b45529
Roller-URL: https://ci.chromium.org/b/8781963278137184993
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I3b5bcc666dd70a99e44e3059964de33e529f8dfa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/852700