[starnix] Only wake on thread in remote_binder
The previous implementation of the waiter in remote_binder always woke
up every thread each time a new command was queued.
In this CL, a single thread is woken up. When the command can be
executed by any thread, a random one is chosen. When it is tailored to a
given thread, the value selection of the WaitQueue is used to ensure the
right thread is woken up.

Change-Id: I0a891eb0b94cd96e9c6d5425f45fc4c695fc1e25
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/852877
Reviewed-by: Adam Barth <abarth@google.com>
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>