[assembly] Only create the archive for the legacy AIB if the product requests it
Only create the legacy AIB for a product when that product requests
it, as opposed to doing so for any release build.  This should save
seconds on smaller builds, and up to a minute on larger builders.

Change-Id: I2dfbb3596343c045b452926a94305c3915c8347f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/852301
Reviewed-by: Aidan Wolter <awolter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>
Reviewed-by: David Iserovich <diserovich@google.com>