[kernel] Fix vmar alignment test flakes
The AlignmentVmar* test cases flake because they try to allocate a VMAR
of size 8GiB with alignment 8GiB, which sometimes is not satisfiable
depending on existing allocations (which are non-deterministic due to
ASLR). While we can't completely eliminate the possibility of this
happening, we can reduce the probability by decreasing the alignment
requirement to 512MiB, as this is enough to satisfy the needs of the test
case but allows for many more available slots and doesn't significantly
increase the runtime of the test.

Bug: 126070
Change-Id: I5e75b360a441538ed037d295a9669cc5713f4983
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/852698
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>