[kernel][vm] Allocate slot before adding page
CloneCowPageLocked migrates or forks the target page down the clone
chain. It is possible for allocation of a slot to fail in the page list
of a child in the chain, which would cause AddPageLocked to fail.
Pre-allocate the slot so that this cannot happen. This avoids having to
roll back in the case where we are migrating the page and are halfway
through, i.e. we've removed the page from the parent but haven't
inserted it in a child yet.

Bug: 126368
Run-All-Tests: true
Change-Id: I6a8cc5a28ccd60cf4282d84ed8d4a48ef47c4cba
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/852258
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>