[pkg] Package manifest should always be loaded non-relative
Developers want to work with paths as relative to CWD instead of
relative to the package manifest itself. When we write a package
manifest as relative to file, we should not return a manifest relative
to the file.

Bug: 126529
Change-Id: Ibd12843f17793e27db4f4d0afeb4ba7f4c35a21c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/851811
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>