[ffx][flash] Also erase misc before rebooting to bootloader
Before this change, when flashing via fastboot in infra, our flashing
flow is:

* unlock the device
* flash the TPL and BL2
* write the GPT
* reboot to the bootloader
* flash zircon + vbmeta slots a, b, and r
* flash fvm
* stage authorized_keys
* erase misc, set slot a active, and attempt to boot

On Nelsons, there are two layers of A/B: one at the TPL layer, and one
at the A/B/R layer.  If the device was in boot slot B before all the
above, then when we reboot to the bootloader, the bootloader is in slot
B, whereas once we erase misc and attempt to boot, that implies booting
from slot A, which is a mismatch, which causes the device to do another
reboot, which succeeds but causes us to lose the staged authorized_keys.

The firmware team recommends doing the "erase misc"/"set slot a active"
steps *before* the "reboot to the bootloader" step above, which should
ensure that we are running the bootloader from slot A, which should
prevent the mismatch.

The firmware team also assures me that erasing misc earlier is safe, and
has the effect of resetting the default boot slot to slot A (for both
TPL and A/B/R) and resetting the unsuccessful boot counts, all of which
are fine in the context of reflashing the bootloader itself.

Bug: 47531
Bug: b/276784189
Change-Id: I69d7d993ec4950e9feac3e638173f0fcd0226c13
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/852636
Reviewed-by: Megan Batty <mgnb@google.com>
Commit-Queue: Drew Fisher <zarvox@google.com>
Reviewed-by: Colin Nelson <colnnelson@google.com>