[build] Move loops over toolchain-static data out of template
The scopes created in these loops in test_spec() are over data that
does not change for a given toolchain, and so should not be re-
calculated for every instantiation of the template (thousands of
times), but only once when the template is defined by the importing
of the gni file within the context of a given toolchain.

This results in no change to tests.json, and removes 55 seconds of
cpu-time from GN gen, reducing the average time per invocation of
this template from 18ms to 2.5ms.

Run-All-Tests: True

Change-Id: I139ba8666be0dd0d4c008ddbcd38370a2ee387c3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/853116
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>
Reviewed-by: Nathan Mulcahey <nmulcahey@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>