[forensics][crasher] Fix RISC-V compilation
Use valid assembly syntax for a known invalid instruction.  The
previous code used ARM syntax not available in RISC-V, and the
value it tried to encode is a newer extension instruction, not
one in the range guaranteed to be an invalid instruction.

Change-Id: Ibe7d47930e870df0e1b436af75933669909abbbf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/853356
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>