[devmgr_config] Rename devmgr_config to additional_boot_args
The old name has a lot of legacy associated with it which no longer
applies. In truth, this config could likely be phased out entirely, but
that is a larger change. It is easier to simply rename it to be more
accurate for now.

Bug: 91451
Change-Id: I774366c015421bbdc7a0476ec34ba5467ee2274b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/842837
Reviewed-by: Laura Peskin <pesk@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>