[kernel][vm] Remove unused all_vmos_cursors_
The all_vmos_cursors_ held any cursors that were iterating over the
all_vmos_ list and could drop the global VMO lock. The only usage of
this used to be for harvesting accessed bits, which was removed in
fxr/533288 that moved all the harvesting logic over to the MMU side.

Even though the only remaining use of the VmoCursor now is for the
discardable list, leave the VmoCursor class templated to allow for more
cursor types in the future.

Change-Id: Ib5b043bbb0e16e9a146c759b1016e70d5e1e1469
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/854544
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>