[build][rbe] Handle special cases with '.'
When build_subdir == '.', there is no relative path to match,
so skip all pre/post-flight checks.
This makes it easier to manually drive remote actions
outside of an `fx build` for testing.

Handle cases when build_subdir contains a '.', which should
be interpreted as a literal '.', not regex for "any-char".

Change-Id: I28de7ed0291435e239a4bd70b5547a3cfb214487
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/854736
Commit-Queue: David Fang <fangism@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>