[bt][a2dp] Restart streams when unpaused
After successfully suspending a stream or not starting the stream
because a permit was not available initially (in contrast to it being
revoked), add a reservation to the queue so that when permit is
available, we will start the stream.

Some minor cleanup of tests as well.

Bug: b/277376263
Bug: 81265
Test: updated tests; fx test bt-a2dp-lib-tests
Test: connect while call in session, finish call and play music
Change-Id: I1a27f77c03014435d0b6c1c19ba73a33ca7c7cbc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/845636
Commit-Queue: Marie Janssen <jamuraa@google.com>
Reviewed-by: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Dayeong Lee <dayeonglee@google.com>