[perfetto-bridge] Destroy ConsumerAdapter before test fixture stops
Lsan on the hwasan FYI builder caught a leak in the
perfetto_unittests_package test. It looks like the test fixture would
close and the program would end (and thus invoke leak checking) before
the actual dtor for the allocation is called in a Post task. So there's
a brief period of time where the allocation is not owner by a
unique_ptr. If the test fixture ends and closes the loop before the Post
resolves, then the pointer is left dangling.

This ensures the test fixture explicitly destroys the allocation before
it closes.

Fixed: 126702

Change-Id: I58518ca10c8378c07711e16116a1fee5cf8a2e3c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/854020
Commit-Queue: Leonard Chan <leonardchan@google.com>
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>