[assembly] Move empty sensor config to internal repo
This include is now included in an internal repository, and no longer
needs to be included here. This transition moves the GN include to the
same place as the product config so that they can be updated together.

Bug: 126530
Change-Id: I02287837cbaeff60b3634e44de4501e65ca0f079
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/855196
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: David Iserovich <diserovich@google.com>