[superproject] Roll antlion Cleanup unused libraries in Antlion
This CL some of the unused SL4F libraries that we do not need because
Antlion is not currently used to test things like Bluetooth. If we
require these functions in the future we can re-add them at that point.

Original-Bug: b/281905321
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/antlion/+/854866
Original-Revision: e52a598c6d1599e036e778f8b81084cb15bb56a7
GitOrigin-RevId: 3f6f044dbdc665ca371a3cb8e58dffe516dbe148
Roller-URL: https://ci.chromium.org/b/8781353194237723265
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: Iaac4cf151b0f9f5b67dc24beee77dcfe2df5e65d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/855460