[virtualization] Implement PL011 LCR
Zircon has started to write to this register which is causing test
failures since the vmm is exiting in response to the unimplemented
register.

This is a trivial implementation that simply makes LCR read/write
without implementing any specific behaviors. This is sufficient to get
the PL011 to work with Zircon guests.

Test: fx test virtualization-core-tests
Change-Id: Idbf5c42d6bafc2b2e56a83a4a43b95d33cfabfd7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/855248
Commit-Queue: Tim Detwiler <tjdetwiler@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>