[ffx][config] Atomically create new user config
In fxrev.dev/853717 I intended to create a 'polyfill' for the lack of
File::create_new but I did it wrong. This makes it so that sometimes a
new config file is created when it shouldn't be.

Change-Id: Ib1668a6419f97b9f5781ba29b7c87e8544e2d4c0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/855462
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>
Reviewed-by: Ina Huh <ihuh@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>