[scrutiny][verify][routes] Correctly handle `--response-level`
`ffx scrutiny verify routes` supports a `--response-level` option, but
it has been handled inconsistently. The option would be passed to the
controller, causing errors, errors and warning, errors and warnings and
capability summaries, or errors and warnings and full route descriptions
to be encoded in the controller's response.

However, the verifier that consumes the controller's response was
ignoring all non-error data, yielding no change in the overall tool's
behaviour when `--resoponse-level` was set.

This change reports a JSON snippet and human-readable messages for all
data returned by the controller. Errors or warnings that are not
filtered out by allowlists result in a non-zero status code. Other
information is printed, but, in the absence of errors or warnings,
yields a status code of 0.

Bug: 103258
Bug: 120171

Change-Id: I09a2302cca51b166517387e9501d3162da1fe77c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/851776
Commit-Queue: Mark Dittmer <markdittmer@google.com>
Reviewed-by: Laura Peskin <pesk@google.com>