[netlink] Support specifying multicast groups
Introduce a pair of clients: the external client will be held by
Starnix, and the internal client will be held by the Netlink async
worker. A client corresponds one to one with a Netlink socket.

Allow the external client to specify the multicast group memberships,
and allow the internal client to observe the current memberships.

Bug: https://issuetracker.google.com/280483454

Change-Id: I15a37956e4f4c3c34ae91b5f775e222ebb9aee3e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/853196
Reviewed-by: Adam Perry <adamperry@google.com>
Reviewed-by: Antonio Linhart <antoniolinhart@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>