[netstack3] Set socket device if addrs require
If the source or destination address of an IP socket have link-local
scope (but are not loopback), set the address of the device in the
socket definition to the egress device. Use this device in TCP and UDP
code to set the bound device for sockets.

Fixed: 125733
Change-Id: Ic95a5a10dfd7ff196af0c31f1e01f632a918438f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/855858
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Alex Konradi <akonradi@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>