[lowpan][openthread_fuchsia] Make receiving TREL packets reentrant.
OpenThread may try to send packets out on the TREL interface directly
from the call that is handling a packet being received on the TREL
interface. This reentrancy was causing the `RefCell` that holds the
TREL implementation to panic. This change makes it so that the `RefCell`
is only borrowed immutably (instead of using `borrow_mut()`) when
processing I/O that might be lead to reentrant calls.

Change-Id: I7f8246adff85ccbee3e85e7993b67647dbcf54ff
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/855824
Commit-Queue: Robert Quattlebaum <rquattle@google.com>
Reviewed-by: Jiaming (Charlie) Wang <jiamingw@google.com>