[netstack3] Simplify TCP payload sending
Previously, the `SendPayload` type implemented the `packet::Serializer`
and `packet::FragmentedBuffer` traits. Those traits (especially the
former) are complicated and not designed to be implemented outside of
the `packet` crate.

In this change, `InnerPacketBuilder` is implemented for `SendPayload`,
which is significantly simpler. `FragmentedBuffer` is no longer needed.
Given `s: SendPayload`, `s.into_serializer()` produces a `Serializer`,
so the `Serializer` impl on `SendPayload` can also be removed.

Change-Id: Ide753123160436f37525fac1fff66e256f607e05
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/856936
Commit-Queue: Joshua Liebow-Feeser <joshlf@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>