[netstack] Support add_subnet_route
Supports the `add_subnet_route` AddressParameter option by adding
the on-link subnet route when the address is added and removing it when
the address is removed.

Note that this approach is not smart at all, and doesn't implement
any reference-counting semantics, so it could unexpectedly clobber
a manually-configured route. Instead, it records whether the subnet
route was _newly_ added by this option; if so, the route is deleted
when the address goes away.

Fixed: 125716
Multiply: interfaces-admin-fidl
Change-Id: I6e4a3c784432f2aaad9e6819afacea0ae3858693
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/855256
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Evan Wildenhain <wildenhain@google.com>
Reviewed-by: Alex Konradi <akonradi@google.com>
Fuchsia-Auto-Submit: Evan Wildenhain <wildenhain@google.com>