[log][validator] Refactor validator into its own package
The goal of this is to have a package that we can use OOT when we move
the Dart libraries OOT. That way we can write the test OOT and use the
validator package built in tree.

This approach relies heavily on subpackages and allows to define a new
test by creating a new puppet and a new package using the test topology.
When moving the dart stuff OOT we won't be able to reuse the test logic
(as we are doing right now), but that should be fine as it's a very tiny
main that leaves all heavy logic to the validator component which we'll
provide to the dart test oot.

I think there's potential for generailization here and integration with
TestSuite. I think this pattern could serve as a based for a generalized
conformance testing framework.

Bug: b/240154194

Change-Id: I67d6c458c112dbeec926ead0b14f62554cc30740
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/855396
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>